Allow for ranging intervals other than 1s

+++ software/firmware/src/tasks/app_tasks.h
@@ -43,7 +43,7 @@ bool ranging_active(void);
 void storage_flush_and_shutdown(void);
 void storage_write_battery_level(uint32_t battery_voltage_mV);
 void storage_write_motion_status(bool in_motion);
-void storage_write_ranging_data(uint32_t timestamp, const uint8_t *ranging_data, uint32_t ranging_data_len);
+void storage_write_ranging_data(float timestamp, const uint8_t *ranging_data, uint32_t ranging_data_len);
 
 // Main Task Functions
 void AppTaskRanging(void *uid);
